requirements.txt
requests
selenium
webdriver-manager
python-dotenv
pandas
numpy
google-generativeai>=0.3.2
scikit-learn
PyPDF2>=3.0.1
pdf2image
pytesseract
pdfplumber
camelot-py
PyMuPDF
pdf2docx
python-docx
openpyxl
python-pptx
Pillow
spacy
InquirerPy>=0.3.4
tabulate>=0.9.0
colorama>=0.4.6
termcolor
asyncio
nest_asyncio
cryptography
qrcode>=7.4.2
pywin32
comtypes
matplotlib
python-dateutil
geminidata
reportlab>=4.1.0